Increase share limit of Luau demo (#58)
## Problem

The current share feature has a hard limit of 2047 characters after gzip and url-encoding. This is due to browser limitations on how long a query param value can be.

Also it doesn't support sharing code with emojis.

## Solution

It turns out that the fragment (#) part of the url is more suited to this purpose than the query param. It actually isn't sent to the server and instead is only parsed on the client. Because of that, browsers don't restrict the size of the fragment portion of the URL to the same extent that they do with query params. Practically, chrome does limit the size of the URL to 2MB so that's what I set as the limit.

Gemini Deep Research actually discovered this for me, although I did write all of the code by hand: https://g.co/gemini/share/9712e01a0afe

Supporting unicode was as simple as processing each character in the array buffer one by one instead of all at once. This is probably slower but sharing doesn't have to be realtime

Example:

I went ahead and copied an entire copy of Frankenstein. Attempting to run the "code" will crash the interpreter but the share function works just fine!

https://github.com/user-attachments/assets/a1c32274-f3e0-48e1-996e-44bade63306e